Make mainnet/testnet switch more obvious Implement react-router-dom for navigation closes #60
