+++ package.json
@@ -6,7 +6,8 @@
     "bulma": "^0.9.4",
     "react": "^17.0.2",
     "react-bulma-components": "^4.1.0",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "react-router-dom": "^6.10.0"
   },
   "scripts": {
     "start": "react-scripts start",
